requirements.txt
fastapi
uvicorn
cryptography
python-dotenv
pydantic-settings